fix 1-12
The program does not take into account if blank, tab or newlines are repeated. 

Example:
test.txt (notice blanks after 'This is a')
```
This is a   succession of blanks.
```

`./a.out < test.txt` 
```
This
is
a


succession
of
blanks.
```

whereas we expect

```
This
is
a
succession
of
blanks.
```














succession
of
blanks.