Update strncmp.c
Increased the accuracy of the comment description, and removed the " != '\0' " part of the while test. Comparing two characters for equality doesn't require seeing if their comparison is not equal to null, it is the same as writing "while ((x == y) != 0)". If x and y are ints or any other datetype, the '==' check suffices when checking for a boolean value.