Fix readlines.c
At the test-part of the if-statement, please recall that we're checking for the *total* number of characters used in storage out of 10,000; we're not checking for the string-length of stored_lines added with the current line length being more than 10,000. Especially when neither is stored_lines itself being changed in the program nor is the length of the current line allowed to be more than 1,000 to begin with.